app/requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==2.5.2
asgiref==3.2.3
asgi-redis
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
Babel==2.8.0
bcrypt==3.1.2
billiard==3.6.1.0
boto==2.49.0
boto3==1.10.47
botocore==1.13.47
celery==4.4.0
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
commonmark==0.9.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
decorator==4.4.1
Django==3.0.2
django-analytical==2.2.1
django-common-helpers==0.9.2
django-cron==0.5.1
docutils==0.15.2
future==0.18.2
graphspace-python==0.9.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.3.0
incremental==17.5.0
Jinja2==2.10.3
jmespath==0.9.4
kombu==4.6.7
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==8.0.2
msgpack==0.6.2
msgpack-python==0.5.6
networkx==1.11
packaging==20.0
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycrypto==2.6.1
Pygments==2.5.2
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==5.3.2
python-dateutil==2.8.1
pytz==2019.3
recommonmark==0.6.0
redis==2.10.6
requests==2.10.0
s3transfer==0.2.1
service-identity==18.1.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.1.5
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.8